requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
colorama==0.4.6
distro==1.9.0
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.11.1
openai==2.7.1
packaging==25.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
